adicionando o menu da calculadora

+++ teste-de-qualidade/calculadora/src/main/java/com/exemplo/calculadora/operacoesmatematicas/Operadores.java
@@ -1,11 +1,13 @@
 package com.exemplo.calculadora.operacoesmatematicas;
 
+import java.math.BigInteger;
+
 public interface Operadores {
     int adicao(int a, int b) throws IllegalArgumentException;
     int subtracao(int a, int b)throws IllegalArgumentException;
     int multiplicacao(int a, int b)throws IllegalArgumentException;
     double divisao(double a, double b) throws IllegalArgumentException;
-    int fatorial(int a) throws IllegalArgumentException;
+    BigInteger fatorial(int a) throws IllegalArgumentException;
     double potencia(long base, long expoente);
 
 }
